+++ src/test/resources/org/csanchez/jenkins/plugins/kubernetes/pipeline/docker.groovy
@@ -1,0 +1,24 @@
+pipeline {
+  agent {
+    kubernetes {
+      label 'docker'
+      containerTemplate {
+        name 'docker'
+        image 'docker:1.11'
+        ttyEnabled true
+        command 'cat'
+      }
+    }
+  }
+  stages {
+    stage('Run maven') {
+        steps {
+            container('docker') {
+                withDockerRegistry(registry: [credentialsId: 'ContainerExecDecoratorPipelineTest-docker']) {
+                    sh 'hostname'
+                }
+            }
+        }
+    }
+  }
+}